+++ tetris/modes/marathon_a3.lua
@@ -9,10 +9,8 @@ local MarathonA3Game = GameMode:extend()
 
 MarathonA3Game.name = "Marathon A3"
 MarathonA3Game.hash = "MarathonA3"
-MarathonA3Game.tagline = "The game gets faster way more quickly! Can you get all the Section COOLs?"
-
-
-
+MarathonA3Game.description = "The game gets faster way more quickly! Can you get all the Section COOLs?"
+MarathonA3Game.tags = {"Marathon", "Arika"}
 
 function MarathonA3Game:new(secret_inputs)
 	MarathonA3Game.super:new(secret_inputs)