+++ tetris/modes/phantom_mania2.lua
@@ -9,10 +9,8 @@ local PhantomMania2Game = GameMode:extend()
 
 PhantomMania2Game.name = "Phantom Mania 2"
 PhantomMania2Game.hash = "PhantomMania2"
-PhantomMania2Game.tagline = "The blocks disappear even faster now! Can you make it to level 1300?"
-
-
-
+PhantomMania2Game.description = "The blocks disappear even faster now! Can you make it to level 1300?"
+PhantomMania2Game.tags = {"Invisible Stack", "Survival", "Gimmick", "Cambridge"}
 
 function PhantomMania2Game:new(secret_inputs)
 	PhantomMania2Game.super:new(secret_inputs)