+++ CHANGELOG.md
@@ -1,3 +1,13 @@
+# v0.4.4.2:
+- (Highscores) frames are now formatted to time
+- Attempted to fix more crashes related to the lack of secret inputs
+- Fixed secret inputs not being passed through on Big A1, Big A2, and Sakura
+- Fixed clearBottomRows function
+- T.B.D
+
+# v0.4.4.1:
+- Fixed bugs related to 21h
+
 # v0.4.4:
 The changes shown here are not in chronological order.
 ## Major changes: